Rename FetchServer to FetchMockServer and Server to SinonServer
