+++ src/FakeRest.ts
@@ -1,9 +1,17 @@
-import { Server } from './Server.js';
-import { FetchServer } from './FetchServer.js';
+import { Server, SinonServer } from './SinonServer.js';
+import { FetchServer, FetchMockServer } from './FetchMockServer.js';
 import { Collection } from './Collection.js';
 import { Single } from './Single.js';
 import { getMswHandlers } from './msw.js';
 import { get } from 'lodash';
 
-export { Server, FetchServer, Collection, Single, getMswHandlers };
+export {
+    Server,
+    SinonServer,
+    FetchServer,
+    FetchMockServer,
+    Collection,
+    Single,
+    getMswHandlers,
+};
 export default { Server, FetchServer, Collection, Single, getMswHandlers };
